+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - `iexec wallet bridge-to-sidechain <nRlcAmount>` sending RLC from a mainchain to the bridged sidechain.
 - `iexec wallet bridge-to-mainchain <nRlcAmount>` sending RLC from a sidechain to the bridged mainchain.
 - brindging RLC between mainchain an sidechain may throw `BridgeError`
+- `deal.claim(dealid)` and `iexec deal claim`to claim all failed of a deal
 
 ### Changed